version-3/requirements.txt
fastapi
fastapi[standard]
uvicorn
sqlalchemy
asyncpg
pydantic
psycopg2-binary
databases
streamlit

